Inline device pool DNS pending logic into DeviceStubResolver
- Remove pending_device_pool_dns.rs and move all pending device pool DNS
  buffering logic into DeviceStubResolver for better encapsulation
- Expose PendingQuery from device_stub_resolver and update all call sites
  to use DeviceStubResolver::{insert_pending, remove_pending}
- Move DNS response construction for device pool queries into
  DeviceStubResolver::build_response
- Remove DEVICE_POOL_DNS_TTL constant from client.rs; use DNS_TTL in
  device_stub_resolver instead
- Update ClientState to no longer store or interact with
  pending_device_pool_dns, relying solely on device_stub_resolver for
  pending and cache management
